+++ test/fixtures/simple.rb
@@ -1,7 +1,17 @@
 # frozen_string_literal: true
 
 module Simple
-  class A
+  class B
+    def self.ffu # rubocop:disable Style/ClassMethodsDefinitions
+      1
+    end
+
+    def bbar
+      2
+    end
+  end
+
+  class A < B
     def self.fu # rubocop:disable Style/ClassMethodsDefinitions
       1
     end